requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.9
charset-normalizer==3.4.2
cssselect==1.3.0
dnspython==2.7.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
lxml==6.0.0
lxml_html_clean==0.4.2
numpy==2.3.1
openpyxl==3.1.5
parse==1.20.2
portalocker==2.10.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyee==11.1.1
pymongo==4.13.2
pyppeteer==2.0.0
pyquery==2.0.1
PyYAML==6.0.2
qdrant-client==1.14.3
requests==2.32.4
requests-html==0.10.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
zipp==3.23.0
